Bump minSdk in one sample to 21
The "container" sample relies on `<tag>` elements in its layout XML to
associate views with their Coordinators.  This sample had its minSdk set
to the project minSdk (15), but `<tag>` didn't appear until 21, so that
was just broken.

This isn't the only (or even necessarily the best) way to associate
views and Coordinators. This sample is just meant to illustrate that
there's more than one way to do it. If you need to support earlier API
levels, you'll need to come up with another mapping scheme.