Fix bind's overeagerness, tighten invariants
After creating a Binding, Coordinators.bind would immediately call its
`onViewAttachedToWindow`, without regard for whether that was true.

The idea was to eliminate a potential race, where `bind` was called
after the View was already attached to the Window. When that happened the
Binding wouldn't be called by the platform, so the Coordinator would
never learn that it was attached unless `bind` called it directly.

That implementation created a problem though: If the View was _not_
already attached to the Window, the Coordinator would still receive an
`attach` call. This violated expectations around the interaction between
Coordinator and View lifecycle. Technically that was fine, because no
promises were made; but it was still surprising.

With this change, `bind` will still call `onAttachedToWindow` but _only_
if the View actually is already attached. This should alleviate any
confusion.

As a side effect, there's also no longer any potential for a duplicate
`onAttachedToWindow` call. This means we can tighten up the assertions
in the Binding. I also took the opportunity to tighten up all of the
assertions in Binding, so now it enforces a very strict 1:1:1
relationship between the Binding, View, and Coordinator and fails loudly
if that invariant is violated.